requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.55.0
django-ckeditor==6.5.1
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-filter==23.2
django-js-asset==2.0.0
django-payments==2.0.0
django-phonenumber-field==7.1.0
django-ranged-response==0.2.0
django-rest-auth==0.9.5
django-simple-captcha==0.5.17
djangorestframework==3.14.0
fontawesomefree==6.4.0
idna==3.4
Markdown==3.4.4
oauthlib==3.2.2
phonenumberslite==8.13.22
Pillow==10.0.0
prettytable==3.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.6
pycparser==2.21
pygame==2.1.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
source==1.2.0
sqlparse==0.4.4
stripe==5.5.0
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uuid==1.30
wcwidth==0.2.5
